requirements.txt
pip~=24.1.1
setuptools~=70.1.1
wheel~=0.43.0
ruamel.yaml~=0.18.6
mkdocs~=1.6.0
mkdocstrings~=0.25.1
mkdocstrings-python~=1.10.5
mkdocs-material~=9.5.27
mkdocs-material-extensions~= 1.3.1
mkdocs-include-markdown-plugin~=6.2.1
pyinstaller~=6.8.0
